Drivers/hv: Widen apic_id parameter of hv_call_add_logical_proc() to u64
The HVCALL_ADD_LOGICAL_PROCESSOR hypercall input (struct
hv_input_add_logical_processor) carries a 64-bit apic_id on ARM64 —
where it holds the MPIDR_EL1 affinity value — and a 32-bit apic_id on x86.
The current hv_call_add_logical_proc() helper only takes a u32, which
silently truncates the ID on ARM64 resulting in hypercalls issued multiple
times for the same ID which the hypervisor doesn't allow.

Widen the apic_id parameter of hv_call_add_logical_proc() and its
prototype in from u32 to u64. On x86 the value is implicitly narrowed
back to u32 when assigned into input->apic_id, so behavior there is
unchanged. Required for ARM64 mshv root partition bring up.

+++ drivers/hv/hv_proc.c
@@ -178,7 +178,7 @@ bool hv_result_needs_memory(u64 status)
 }
 EXPORT_SYMBOL_GPL(hv_result_needs_memory);
 
-int hv_call_add_logical_proc(int node, u32 lp_index, u32 apic_id)
+int hv_call_add_logical_proc(int node, u32 lp_index, u64 apic_id)
 {
 	struct hv_input_add_logical_processor *input;
 	struct hv_output_add_logical_processor *output;
@@ -207,7 +207,7 @@ int hv_call_add_logical_proc(int node, u32 lp_index, u32 apic_id)
 
 		if (!hv_result_needs_memory(status)) {
 			if (!hv_result_success(status)) {
-				hv_status_err(status, "cpu %u apic ID: %u\n",
+				hv_status_err(status, "cpu %u apic ID: %llu\n",
 					      lp_index, apic_id);
 				ret = hv_result_to_errno(status);
 			}
