+++ include/asm-generic/mshyperv.h
@@ -419,7 +419,7 @@ static inline bool hv_parent_partition(void)
 
 bool hv_result_needs_memory(u64 status);
 int hv_deposit_memory_node(int node, u64 partition_id, u64 status);
-int hv_call_add_logical_proc(int node, u32 lp_index, u32 acpi_id);
+int hv_call_add_logical_proc(int node, u32 lp_index, u64 acpi_id);
 int hv_call_create_vp(int node, u64 partition_id, u32 vp_index, u32 flags);
 int hv_call_notify_all_processors_started(void);
 u64 mshv_current_partid(void);
@@ -435,7 +435,7 @@ static inline int hv_deposit_memory_node(int node, u64 partition_id, u64 status)
 {
 	return -EOPNOTSUPP;
 }
-static inline int hv_call_add_logical_proc(int node, u32 lp_index, u32 acpi_id)
+static inline int hv_call_add_logical_proc(int node, u32 lp_index, u64 acpi_id)
 {
 	return -EOPNOTSUPP;
 }
